auth/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore
google-cloud-pubsub
google-cloud-tasks
requests
PyCryptodome
